Use json.Unmarshal to write to a strange type in v1 fallback test
The result value for the watcher API call in
TestWatchRelationChangesV1Fallback couldn't be assigned to in the same
way as other API call results. I think this is because of the way the
commonWatcher code gets the result from a func that returns
interface{} and then passes it into the call as &result, so
it's *interface{}, which seems to cause problems. The RPC call can
write into it fine using json.Unmarshal, so change the test to use
that as well.

Also noticed that the watchers could be stopped as far as the tomb was
concerned before the commonLoop had finished (because Stop wasn't
being called), so changed them to start commonLoop inside the tomb
instead.