Deploy tmp directory
The following changes ensure that we create a new unique tmp directory
that we can deploy from. The previous code created this tmp directory
but so did the library function. Assuming we actually want to create a
tmp directory to launch from, I've ensured that we namespace a new
directory inside the tmp directory that the library can create. Although
convoluted, it ensures that we get what we want; ephemeral directory
state and the library gets to create it's own directories that it can
own and manage accordingly.