Handle nil body requests
The following handles nil body requests when serving stub integration
calls. This does show that having stub code means it's really difficult
to mirror code found with in the stdlib in its entirety. This would be
probably better served as mocks with additional integration tests
preformed on the client directly with stubbed API to gain isolation.