Use the multiwatcher worker in the apiserver and model cache.
This requires updates to the manifolds. For the model cache it replaces
the need to get the state pool. For the apiserver, it needs the
multiwatcher worker in addition to its other pieces.

The multiwatcher worker is now started in the apiserver tests and
in the dummy provider to make the multiwatchers available.