Build less version checking
The following brings in build less version checking. When looking for
streams or updates, using the build numbers to signify a change isn't
required. The uniqueness of a build number offers to much variation for
juju to check, instead Major.Minor.Patch.Tag is enough as that's
currently what we store for streams.

These changes should fix the issues around not being able to bootstrap
without the remote source available if you're not on ubuntu.