Backwards compatibility
The following updates show-storage so that it's backwards compatible
with the feature tests around it. Essentially the fact that the user
asked for formatted directive of yaml and instead was returned a string
doesn't match what they asked for.

The fix here is to say, return an empty object `{}` (remembering that
yaml is a subset of json) and then correctly outputting the issue to
stderr.

The consumer of the CLI can then either check stderr for issues or
redirect stderr to a file or /dev/null. Either way the consumer gets
valid yaml and you also get the correct information about why it failed.

See: https://discourse.jujucharms.com/t/cli-serialisable-output-formats/2490