Removes check in migration export that skips relation unit settings for cross-model relations.
This check was causing the inference of relation scopes to fail on the
import side because *no* relation-unit settings were present for such
endpoints.

Now they are present for *local* units participating in the relation.

The check was not required anyway, because there will be no units for
remote applications in the local model.