Get CharmProfilingInfo from state
The following changes are a theory that we're requesting
SetCharmProfiles too many times. This leads us to believe that before
calling SetCharmProfiles, we ask what does the current charm profiling
info look like? If the charm profiling info matches what we want to set,
we return out and exit early. The flame graph of the goroutines show
that the API is call thousands of times, but nothing further down in the
controller (state) gets called. The txn for the operation exits out as
it does another check from the db itself as a no operations error.

So how are we getting to the SetCharmProfiles considering the db has the
information. Under the hood, CharmProfilingInfo uses the cache, is it
that the cache isn't populated in time, esp. in HA mode?

Can it be then that using the state might prevent this? This PR aims to
solve that by changing the dependency from the cache to state.

How we test this at scale is a concern though...