Prevent export bundle trimming with certs
The following changes ensure that we preserve the export bundle output
even if there is content in there that could cause issues with overlays.

The overlay code checked for dashes to exist and if found, trim the
content and add a comment, before adding back the overlay. This worked
fine unless you had a cert in your options for an application. Doing so
caused conflict and made the yaml appear cut off.

The fix was to encode the two structures independently and then stitch
them together with the comment inbetween.

Tests have been introduced to ensure that a CA cert can be dumped in any
part of the bundle and that the yaml is correctly outputted.