Handle detaching a volume attachment with a plan.
If the VolumeAttachmentPlan is already in Dying, treat DetachVolume as a
no-op.
Because if you *don't* have a plan, and VolumeAttachment is already
Dying, DetachVolume was already a no-op. VolumeAttachmentPlan exists to
get a sync from the machine agent that it needs to run a local command,
and then progress the system by calling RemoveVolumeAttachmentPlan.
If we have already progressed the Plan to Dying, there is nothing else
to do.

This doesn't handle the 'juju destroy-machine --force' case, where the
remote machine will never call RemoveVolumeAttachmentPlan on its own,
but it does prevent us from getting 'state changing too quickly'.