Changes the location of systemd unit files and start-up scripts.
Instead of writing each service definition and it's exec-start script
into a service folder in /lib/systemd/system and using dbus (systemd) to
link them into /etc/systemd/system, we write them directly to
/etc/systemd/system.

This change was required to support Focal Fossa (20.04) which has these
two directories on different partitions, causing systemd to fail.

The linking step remains for now, though it effectively does nothing -
the service definitions are in the default search path, so reloading the
systemd daemon should be sufficient to have them loaded.