Implement state-delete, update state-[get|set] hook tools.
Rename StateValue* to StateCache*, state-value-[get|set] to
state-[get|set].

state-set should not delete keys.  Let's not continue the precendent
set by other set commands in juju.  It prevents a key having an empty
string value and is not intuitive.

Implement juju tool for deleting a value from the charm state for a unit

Add unit tests with mocks for all.