Ensure that wait for consuming model
The following additional wait is to ensure that the consuming relation
is setup before attempting to migrate it. If you try and do it too
quickly then the migration will fail.

We could put a sleep in here, but actually it's better to test the full
flow of communication as that will be inherently much quicker in the
long run, with less failure cases.

The only outstanding issue now is that if this does fail, then
destroying a controller fails because there are still offers in
existence.

This causes dangling jujus in the testing infrastructure. I'll solve
that in another PR.