Clean up offers
During the cleaning up of integration tests, where the test "framework"
destroys the controllers, if there are any outstanding offers in the
controller then the destroying of the controller will hang for ever!

The fix is quite simple, before the actual destruction of a model, walk
over all the controllers and remove all the offers. It's quite a slow
process esp. with AWS, but it does mean that we don't have any orphaned
controllers.

In fact this is how a normal operator is supposed to handle a controller
destruction.