Handle applications being offered with a different name
The appdata changes to the crossmodelrelations and remoterelations
facades incorrectly assumed that the application token always
translates to the application tag, when in fact in offering models it
contains the offer name (as an application tag). This meant that if an
application was offered under a different name the crossmodel code
would return a NotFound error when the worker in the consuming model
tried to watch the new relation, so the worker would mark the SAAS
terminated in its model and give up.

Split GetRelationTokens into a consuming version (which gets the
application name from the relation) and an offering one that gets it
from the offer.