Refactor Open/ClosePort into a ModelOperation
Originally, the open and close port implementations were separate.
Unfortunately, this made their composition virtually impossible given
that both manipulated the list of port ranges in the buildTxn callback
and then proceeded to mutate the in-memory model representation after
the txn completed successfully.

As a result, if we composed their individual []txn.Op together and
executed them in the batch we might end up with inconsistent data in the
state.

For instance, if we have already opened ports 100-200 and then try to
compose {Open(500, 600), Close(100-200)}, the close txn would not have
access to the open side-effects (a new PortRange is appended) and would
assume that the document is now empty (as it only sees the existing
100-200 range in the doc) and delete it!

Similar problems stem from the fact that the close txn uses "$set" to
overwrite the port range list.

This commit addresses this problem by combining the two operations in a
single OpenClosePorts call which invokes OpenClosePortsOperation to
obtain and apply a suitable ModelOperation that updates the state in a
consistent way.

A side-effect of this refactoring is that we can now safely get rid of
all the port-related method variants on the unit model (as suggested by
the TODOs in the code) and replace them with a single
OpenClosePortsInSubnet method which is invoked by the facade code. In
the future, we should refactor this method to work with endpoints
instead of subnet IDs but this is currently outside the scope of this
piece of work.