Remove CMR guard
Now that CMR for remote applications has landed in develop and we have
integration tests for this, we can finally remove the feature flag.

This means that moving remote applications should be now possible. If
there are any issues then report a bug to https://bugs.launchpad.net/juju/+bugs