Replace SettingsGroup with composed ModelOperations
The Settings model has been enchanced with a WriteOperation method which
returns a ModelOperation for persisting mutations to the settings model.
The UpdateNetworkInfo API method in the uniter facade leverages this new
call to compose a sequence of WriteOperations to update the network info
(ingress, egress) for joined relations (each relation has its own
Settings instance) and execute them in a single transaction.

To this end, the SettingsGroup type is now made redundant and removed as
part of this commit.