SubnetID for openstack networks
The following adds a subnet which is derived from a AZ, on to a
nova.ServerNetworks. It uses the fixedIP field (should we rename this?)
as we believe that fixedIP can be used as a subnet as well.

The code pretends to be networks (plural) aware, but in reality it is
only singular. The DefaultNetworks returns an empty slice and the
ResolveNetwork returns one item. It currently guards against finding
multiple networks, so the current code should be fine. Overlapping
subnets on multiple networks will be intersting!

This code is mainly a rip off from the ec2, but I'm unsure if it's
correct, so please correct my naivety.