Attempt to select a CIDR
The following selects a CIDR from the network using a UnknownId for the
instance selection. I believe the neutron networking only provides
support for this, which is exactly what we want as we don't know the
instance id at this point. So we get all the available subnet ids for
the network id and we select a CIDR to try.

Because we don't really support multiple networks yet, this code is
fine, future code will have to be adapted to ensure that we use the
right CIDR for the right network.

Future PR will ensure that the FindSubnetIDSForAvailabilityZone returns
a network.Id, so we don't have to jump between types.