Implement server-side logic for the CommitHookChanges API call
This new call essentially executes one or more (depending on whether the
appropriate field in the CommitHookChangesArg is populated by the
caller) of the following API calls in a single transaction:

- UpdateNetworkInfo: invoked after a config-change event to sync network
information for each joined relation in case the space assignments have
changed.
- UpdateSettings: updates the relation settings for the unit and/or
application (leader only)
- OpenPorts: open a port range
- ClosePorts: close a port range
- SetState: update server-side unit state for a charm

The call performs the required access checks for the unit once and then
invokes the appropriate methods to obtain a ModelOperation for each
individual API call to be executed. The resulting ModelOperations are
composed together and executed in a single transaction.