Refactor context flush logic to use new call and simplify error handling
This commit refactors the hook context flush logic so that it invokes
the CommitHookChanges API calls for the 5 calls currently
supported.

The AddUnitStorage/PublishPodSpec calls are a bit more convoluted and
require additional work to add them to the consolidate API call. They
will be added using a separate commit.

As part of the refactoring work, the Flush method has been updated so
that error handling is more straightforward as opposed to passing
pointers to named return error values in deferred method calls (e.g.
the call to handleReboot). In addition, the flush logic will bail out as
soon as it gets back an error instead of trying to execute all API
calls. This change makes the flush call's behavior congruent to the
contract that hooks can fail (and be subsequently retried) as they are
expected to be idempotent.