Remove SetState API call
This call was introduced for mutating the unit's charm state that is
persisted by the controller. Now that the hook context Flush() method
has been updated to use the CommitHookChanges API call, there is no
point in maintaining this call. As nothing out there currently uses it,
it's safe to simply go ahead and remove it.