Typed network.Id argument
The following changes push through the idea of the type annotation for
the network.Id through the provider so we don't have to guess what is
being passed back.

This makes the code more concise and more readable imo.

Additionally the change pervents string -> Id -> string and
Id -> string -> Id by creating a new set type.