Introduce a new watcher to the cache to feed the new JAAS dashboard.
In order to test the new watcher, the BaseSuite was moved into the
package_test file. The controller tests were internal, so the methods
needed to be renamed to export them to make them available to the
test package.

The hash function is updated to make the settings map optional.