apiserver subnets: add mocks; status test: skip flaky test to handle later at a sprint
this patch adds mocks to subnets as a preparation for further adding features to subnets.
Additionally, we plan to move from stubs to mocks. This patch should prepare most of the things for it.

The other file touched is the status_internal_test.go where I needed to skip some flaky test because of the possible slow modelcache. We should look into this further at the sprint