Extract open/close port assertions in a helper mixin
To avoid unnecessary repetition of the open/close port assertion logic,
the relevant code has been extracted into a helper that can be embedded
in our test suites. The new helper, FirewallHelper lives in the
core/network/testing package.

The original signature for unit.OpenClosePortsInSubnet proved to be
challenging for this refactoring as the use of state.PortRange arguments
prevented the use of the mixin in some of the internal state tests due to
import cycles (state imports network/testing that imports state). To
this end, the OpenClosePortsInSubnet method has been modified to accept
network.PortRange arguments and convert them internally to
state.PortRange instances.