Initial reworking of networkInstances
Cleaning up the code to make it much more cleaner about where the
networkID comes from. We shouldn't code to the future and state that
there might be some random "DefaultNetworks", when in fact they're is
only ever one.
Trying to map from a networkID to a subnetID without any real
relationship and without brute forcing it by iterating over every port
id looking for a network id, is beyond the pale.

Instead the code says there is only ever one network ID and that
networkID is the one we're going to use to map to every nic, space and
subnet.

There are no tests here, they'll come later.