Validate leadership when updating PodSpecs
This commit enforces leadership checks on the server when attempting to
update the PodSpec for an application. Prior to this commit, this check
was only performed client-side by the uniter agent when the hook context
was being flushed.

Besides the uniter, the same code path for updating the spec is also
used:
- during model migrations
- by the CAAS operator

For the above two cases, the leadership check is bypassed as it is not
relevant to the sub-system performing the call.

This commit fixes LP1863155.