Define a FanType
The following defines a FanType (InFan) and a filtering function for
removing them from a series of networks.

This is required for the openstack provider when we have a space, but we
don't want to pick a random address that could land on an external
network. This check removes those by checking for the FanType with in
the network.Id.