Terminate instance network ports
When creating an openstack instance with spaces, we have to create a
series of ports so that we can land that instance in the right subnet.

To prevent orphan network ports in the operators openstack setup, when
terminating the instance, we go through and ask for all the ports
associated with openstack via the device_id and attempt to delete them
if required.

The code ensures that we don't block when terminating the ports, but
instead report out to the operator via Debug messages that a failure
occured.

In an ideal world we should probably ensure that we report every port id
failure and get them to do it manually, but this could end up being
potentially a lot of information dumped to the operator.