Testing termination of ports
To correctly exercise the termination of ports with in the open stack
provider we need to force the assignment of the device id of the port in
question. The feels a bit wrong, as we're reaching into the the bowls of
the openstack testing framework, but unfortunately as this code is
blackbox rather than whitebox testing, we have no choice.

This should be moved to mocks in the very long term.