Retry machine start for AZs with constraints
This patch fixes lp1860083 where the provisioner would
not retry starting a machine in an AZ when constraints
that limited the AZs the machine should run on.

This patch also fixes an issue with instance type
selection on AWS, where despite the AWS price list
having an entry for an instance-type, that instance
type may be only supported on a subset of AZs in that
region. We compile a list of instance types supported
in AZs, and where an instance-type is not supported in
all AZs, it is marked as deprecated. Users may still
the instance type, but it will require an explicit
constraint.