Dump test failures to the console
Some people have wanted the dump from the test run when there is a
failure on to stdout/stderr.

This is actually really simple, we stream the current test run to a log
file and then when it fails we just cat that to stdout. Nothing
complicated, but really handy for when parsing the logs.

To keep backwards compatibility with what we've currently got, I've
added a new flag -t to turn this mode on. Locally it makes sense to not
have this running, nicer output and you have the files anyway. For
jenkins setups this helps when the output tar isn't created.

I'll modify jenkins to correctly use this flag if it's available.