This patch fixes an a concurrency error due to map write+iteration in the model cache.
The model's updateSummary method was being called by the controller,
unguarded by the model's lock. This means that processing any deltas at
the same time as Sweep is called, could cause a panic.

Testing this fix outed another concurrency bug with passing the
initializing boolean reference to the models where it escaped the
controller's lock and could potentially be read while the controller was
setting it.

The boolean is now replaced by a channel to achieve the same result.