Microstack add space
The following adds setup code for creating a setup for microstack. It
tries to be idempotent in the way it sets up the metadata and the
simplestreams. It's not always going to be right and it puts a big onus
on the destroy working effectively.

Anyway the add space is using the wrong cidr, but we've not got that far
yet!