Introduce helper for managing transient files
The helper implementation is trivial for *nix targets; the caller needs
to pass a directory that will be automatically purged after a reboot
(so anything mounted as tmpfs or a ramdisk e.g. /var/run).

The windows version relies on the MoveFileEx (windows "syscall") exposed
by golang.org/x/sys to schedule the removal of each created transient
file when the system reboots. This is achieved by specifying a nil value
as the destination for the move operation in conjunction with the
MOVEFILE_DELAY_UNTIL_REBOOT flag. For more info see
https://docs.microsoft.com/en-us/windows/win32/api/winbase/nf-winbase-movefileexa