Remove redundant support for SetPodSpec in uniter client
Given that the uniter is now invoking the CommitHookChanges API call,
we don't need to support SetPodSpec in the uniter client as it is
essentially guaranteed that the controller will *always* run the same or
newer juju version. Note that we still need to preserve the SetPodSpec
call implementation in the uniter facade for backwards compatibility
with older uniter versions.