Synchronise while collecting cache metrics.
It has been observed in the wild that the metrics that were being
collected from the model cache was not consistent. The collector
resets the guages and then interatively increments the counters
based on the life and status of the values. If we don't protect
against concurrent collect calls, overlapping calls will cause
under counting in one return and over counting in the other.