Add a failing test for the unit scaling.
Over time the unit processing in status has regressed to where additional
database queries have been added per unit.

The unit test added starts with a single unit of an application on a machine
and records the number of database queries that were used to run full status.
When we create new units, we make sure that they for the same application and
on the same machine, as right now, adding an addtional machine will make
the query count go up as we are doing additional things per machine right
now. We'll fix that one later.