Fix 1860083 suitable availability zone for machine not found
Minor refactor to cleanup branch logic for distribution groups
for AZs.

The actual fix was in dropping the FilterZones method as it was
relying on improper slice expectations. The append mutates the
array, the correct way would have been a nil array.
Go only returns a new underlying array for an append operation
when the array capacity is increased.
See https://play.golang.org/p/gUuDY0y1sqa for an example.