Try adding checks for when we don't have stateInfo.
The only things that need stateInfo are machines with JobManageState, so
the Provisioner shouldn't need it. And we only need a MachineTag if we
are the NewContainerWorker not if we are a NewEnvironWorker, so try to
trap failures when Config doesn't have a valid MachineTag.

So far, it seems to just cause the Controller to restart uncontrollably
(not even logging failure messages.) Still debugging.