Speed up the processing of machines in juju status.
This branch addresses the multiple database queries that were being
run for every machine in the generation of the full status results.

The constraints were being read individually, so we read them in bulk
at the start of the function. State tests were augmented to show that
constraints set on a machine can be read from all the constraints.

Additionally the instance data was being queries three times for
every machine. Once to get the instance id and display name,
another time for the hardware characteristics, and a third time
for the charm profiles. The instance data is now read in bulk at
the start of status and used when processing the machines.

Without these changes the test results showed:
   c.Check(tracker.ReadCount(), gc.Equals, queryCount,
        gc.Commentf("if the query count is not the same, there has been a regression "+
            "in the processing of machines, please fix it"))
... obtained int = 41
... expected int = 21

Which matches expectations, as five machines were added, and the
processing of each added another four queries.