Add commit hook tool use to ubuntu-plus charm.
Updates to use new hook tools for juju in a charm to save
charm data in state, instead of a local db.

The no-dispatch action will move the dispatch file to a
new name.  The dispatch action will move a given file to
dispatch.  Both will delete the contents of charm data.

The update-status hook will set charm data one=two and
commit those changes.  It then sets charm data three=four
and calls the dispatch file.  If the file does not exist
and the hook fails, three=four will not be saved.