Fix a watcher race that was in the ControllerAddressesSuite.
It looks like the two WatchAPIHostPorts tests were moved from state_test.go
around nine months ago, but somehow we ended up with both. I have removed
the copy that was in the state_test file.

The StateSuite had the WaitForModelWatchersIdle at the end of the setup,
but the ControllerAddressesSuite didn't. This allowed for initial creation
events to still be propagating through the system when the tests started.