Allow custom image streams to override force.
The following image streams changes allows a user to specify a image
stream that isn't "released", so that if they're testing a new LTS, they
don't have to supply force.
The idea being, that Juju should be less hostile to new LTS support and
assume the operator knows that they're doing when providing custom
image-metadata-urls in co-ordination with image-streams.

Luckily in bootstrap, force is implied to be - force series - and no
other side effect it means we can be certain about the consequences.
Future commits around deploying a charm and determining this behaviour
might not be so simplistic.