Allow force with deploy bundles
The following attempts to allow the operator to use force with deploying
charms and bundles.

I believe this code to be highly dangerous!

As there is no idea of force series vs force lxd profiles or force
trust (etc) in the deployment of charms. Independantly stating that if
you have a custom image stream of something other than "released" and an
image metadata url other than an offical one, means that you suddenly
have the power to force without any explicit supplied parameter is at
**best** a hidden disaster waiting to happen.

Forcing the machine OS isn't the same as force series, although they're
congruent they're not exactly the same. We should be careful about
entwining them even more.

Conflicting more is that is purely client side and attempting to do this
server side isn't any better. Again we don't have enough context about
what force means (unlike upgrade-charm), did the operator want to force
a lxd-profile or force a trust setting for a bundle. I'm unsure they
wanted to have force rights just be changing a image-stream model
config!

---

Instead we should prise apart the means of force in the command and only
allowing force series to be overridden and not the other types of force
that's implied in deploy.

---

Aside from updating all the permutations for deploying at the last
minute seems like a disaster waiting to happen, esp. considering this
supposedly our last 2.7.x (stable) release.

We should think about this very carefully as it seems like a blanket
rule that's hidden to most isn't very clear.