Series info for a juju release
The following reworks all the series information for a juju release. It
takes the os/series and repackages them from global level package
functions to types that we can reason about.

Adding types allows us to use unit testing along with mocks, as we're
not having to create an elaborate testing infrastructure. We should be
able to pass this through to other call sites if required, rather than
currently executing unfront.

---

The implementation splits the current os/series into two definitive
types, SupportedInfo and DistroInfo. The DistroInfo deals with the hot
reload of the ubuntu.csv that's found on every distro. It has a view of
what's in that csv and nothing more.

The SupportedInfo has the idea of a preset default series it knows about
and using the DistroInfo to load up it's view of the world and update
it's known presets.

The code now has two clear boundaries that than splurged everywhere.

---

Following PRs will deal with two things...

 1. Dealing with non-linux DistroInfo
 2. Loading this into commands