IgnoreDistroInfo to override supported.
Distro source information can force a supported series to be unsupported
because a series might not be released yet. Using OverrideSupported to
force the ignoring of distro info means that we can use the default
value of supported even after a distro source update.

This is required to support focal before it's released, but also applies
to all future releases.

f