Series and daily image-stream logic
The following updates the bootstrap and deploy logic to take into
account the daily image-stream model config. If the series passed in is
not empty and you're using a daily image-stream setting, then attempt to
set the series to be supported.

Using that new logic we can then tell bootstrap and deploy to correctly
check for new releases and if they're found and match a supported
series, go and use that.

This works great for focal.

---

The code also removes the distro info and puts it back into os/series
and in turn makes that available for os/series to use.

The reason for this is to start moving towards a better os/series
package, one that doesn't use package level functions and mutexes
everywhere.

It's non-idiomatic go code, never mind that it's not easy to test.

---

Once or "if" that code lands, we can update the deps to point at the
landed juju and use that.