Address duplication of code issue
In order to make it easier to reason about, I've moved everything that
is duplicated between deploy and bootstrap to core/series. That way we
can conform to a type and have different implementations; controller vs
workload.

The benefit is that we can standardise on set.Strings rather than
set.Strings and []string.

---

Updated the tests for the fallout in tests surrounding the series
selector.
The code now removed some loops which we can use set.Strings for.