Enhance UnitState and add SetState.
Add UniterState, RelationState, and StorageState to unit state saved on
the controller in prepartion to move the uniter internal state from a
unit local files to the controller.

Implement SetState to save the data outside of CommitHookChanges.

Must handle nil vs no value when setting unit state.  Each piece of
the unit state can be changed independently. Changing one piece should
not impact another unless intended. For each value in the unit state,
if it's not set, don't change the current values.  If it has no value,
remove it from the unit state. Other wise update the unit state.