Rename relation resolver state tracker and allow it to be shared
In order to be able to inject relation-created hooks before the
leadership hooks fire, we need a separate resolver just for the
relation-created hooks.

However, both resolvers need to refer to the same backing state since
they need to access information about the currently joined relations and
the Relationer instances that wrap each relation. Unfortunately, this is
not possible with the current implementation as the state implementation
details are abstracted behind a convoluted interface which also includes
methods for preparing/committing hooks and for figuring out the next
hook to execute. In fact, one of the key reasons why we cannot share the
state is that the actual logic for selecting the next hook lives in the
state itself! The resolver's NextOp implementation simply delegates to
the state and generates an operation from the returned hook (if any).

In an attempt to make the state object reusable, this commit effectively
removes the next hook logic from the state and moves it into the
resolver. The state object has been augmented with additional methods to
allow the resolver to make any required checks as part of its next hook
logic (IsKnown, IsImplicit, HasContainerScope, StateDir etc.).

Additionally, to make the code easier to grok and to better indicate the
real purpose of the relation state helper, it has been renamed to
RelationStateTracker and moved to the statetracker.go file. The resolver
has also been moved to resolver.go.

Finally, additional tests have been added to cover code paths that were
not previously exercised (e.g. destroy subordinates when the principal
dies).