Introduce resolver for emitting relation-created hooks
This resolver runs *before* the resolver for leadership hooks and shares
the same StateTracker with the regular relation resolver. Its purpose is
to enumerate the relations appearing on the remote state representation
and emit relation-created hooks for any newly-established relation.

To avoid spurious relation-created events when the uniter agent
restarts, the relation state tracker keeps track of whether
relation-created events have already been fired. Upon initialization,
the state tracker will iterate the remotely known relations and check
whether the local state indicates that we have entered the scope for
each relation. If that's the case, the state tracker assumes that the
relation-created event must have already been fired sometime in the past
(otherwise the relation would not have been joined yet) and mark it as
emitted.