Handle no distro file
Handle the fact that not all ubuntu systems have the distro file, this
is a bit strange for me as that means they have never run a distro
update, esp. when the release series is trusty.

The code brings in the latest os/series package that verifies that the
distro file exists, otherwise return out early.