Introduce a few new fields to the core/multiwatcher types.
The preferred addresses are added to the machine info so as to
make the initial loading of the units not require extra database
queries. These fields will be populated when the machine is loaded
and used when processing non-CAAS units.

CAAS units are another kettle of fish and will require additional
work.