Update the allwatcher to use the new mutliwatcher fields.
Instead of asking the units for their preferred public and private
address, for VM models, we get this information directly from the
machine that the unit is running on.

Also a drive-by removal of a duplicate import and cleaning up of getting
the model from the store.

The special casing of determining the assigned machine ID for subordinate
units is removed based on upcoming work to denormalise the information
and store on the subordinate unit doc.