Fix the tests around subordinate creation.
Many of the tests were not ensuring that the principal units
were assigned before creating the subordinate unit. For those
tests where it matters, the tests are updated to ensure that
the principals have been assigned properly.

Some of the tests were doing icky things, like entering scope
to create the subordinate, then leaving scope again, to make
it seem like we hadn't done things yet. This icky behaviour
is still there, and still icky.