Make sure that models being imported set the subordinate machine ID.
The exporter has a drive by because there was already a database model
available.

The importer will use the machine ID if one is set, as the exporter does
export it if it is set, but if the model is a IAAS model and the machine
ID is missing for the subordinate, then get the machine ID from the
principal unit.