Remove the update space command
The update-space command was originally implemented for the MVP, but was
never implemented at the api client or api facade level. This means if
you actually tried to use it, it would panic for method not found as no
client ever implemented the UpdateSpace method.

Removing this also helps when we want to implement MoveToSpace as there
seemed to be some overlap between the ideas of the two. Instead we will
opt for a more pure MoveCommand.

I believe there could be more clean up as I go, but this is the first
stab at that.