Remove obsolete commands
Be more ruthless and remove all commands that are obsolete.

There is a question around the api/subnets API.CreateSubnet, where we
want to be able to remove a method, but unsure because of backwards
compatibility.

Any resort deleting code feels good as it's code that isn't used, never
tested and just panics when used!