Query local unit/app life and populate DepartingUnit entry in hook.Info
To figure out whether it's the local or remote unit that goes away, the
resolver queries the life value for the local unit and its part app and
checks whether both are alive or any shows up as dying. Depending on the
outcome of this comparison, it populates the DepartingUnit field of the
emitted hook.Info value with the appropriate unit name.

Since the local unit state may be stale, the StateTracker implementation
will first refresh the unit prior to obtain the latest life value as
recorded by the controller.