Expose the CIDR when moving a subnet
We need the CIDR in the cli to show the operator when a move is
completed so they can reason about what has moved.

Simply exposing the CIDR and adding a test for this was easy enough and
it removes the hacks in the client that would have been needed (tracking
what space tag relates to a cidr).