Initial workings of the move-to-space command
There are a few files changed here to add the move to spaces command,
that's because we no longer are talking to 1 api client (facade).
Instead we're talking to SpaceAPI and SubnetAPI. The latter is used to
get a subnet tag from a CIDR.

I think it's not a bad change and it works quite well.

More tests will follow around the move command.