Validate tags for output
Ensure that the tags we get from the api server is valid space tags. We
can then output the name of the space that's been moved to the operator.

The code is rather mechanical, but works for what we need and require.

---

Additionally we also output the command via main so you can use it.